revocation: Extract `crl` submodles to their own files.
